+++ app/Http/Controllers/Admin/JobCategoryController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Job;
 use App\Models\JobCategory;
 use Illuminate\Http\Request;
 use App\Services\Notify;
@@ -76,7 +77,13 @@ class JobCategoryController extends Controller
 
     public function destroy(string $id)
     {
-       
+
+        // validation
+        $jobExist = Job::where('job_category_id', $id)->exists();
+        if($jobExist) {
+            return response(['message' => 'This item is already been used can\'t delete!'], 500);
+        }
+
 
         try {
             JobCategory::findOrFail($id)->delete();
