+++ app/Http/Controllers/Admin/JobRolesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Job;
 use Illuminate\Http\Request;
 use App\Models\JobRole;
 use App\Services\Notify;
@@ -86,6 +87,14 @@ class JobRolesController extends Controller
      */
     public function destroy(string $id)
     {
+
+         // validation
+         $jobExist = Job::where('job_role_id', $id)->exists();
+
+         if($jobExist) {
+             return response(['message' => 'This item is already been used can\'t delete!'], 500);
+         }
+
         try {
             JobRole::findOrFail($id)->delete();
             Notify::deletedNotification();
