+++ app/Http/Controllers/Admin/JobTypesController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Job;
 use Illuminate\Http\Request;
 use App\Models\JobType;
 use App\Services\Notify;
@@ -89,6 +90,14 @@ class JobTypesController extends Controller
      */
     public function destroy(string $id)
     {
+
+         // validation
+         $jobExist = Job::where('job_type_id', $id)->exists();
+
+         if($jobExist) {
+             return response(['message' => 'This item is already been used can\'t delete!'], 500);
+         }
+
 
         try {
             JobType::findOrFail($id)->delete();
