+++ app/Http/Controllers/Admin/ProfessionController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\admin;
 use App\Http\Controllers\Controller;
+use App\Models\Candidate;
 use App\Models\Profession;
 use Illuminate\Http\Request;
 use App\Traits\Searchable;
@@ -87,6 +88,14 @@ class ProfessionController extends Controller
      */
     public function destroy(string $id)
     {
+
+         // validation
+         $candidateExist = Candidate::where('profession_id', $id)->exists();
+
+         if($candidateExist) {
+             return response(['message' => 'This item is already been used can\'t delete!'], 500);
+         }
+
         try{
             Profession::findOrFail($id)->delete();
             Notify::deletedNotification();
